requirements.txt
annotated-types==0.7.0
iniconfig==2.0.0
loguru==0.7.3
packaging==24.2
pluggy==1.5.0
pydantic==2.10.4
pydantic_core==2.27.2
pytest==8.3.4
pytz==2024.2
PyYAML==6.0.2
singleton-decorator==1.0.0
typing_extensions==4.12.2
